+++ core/gui/src/app/workspace/component/property-editor/property-editor.component.html
@@ -1,6 +1,6 @@
 <ul
   nz-menu
-  id="property-buttons1"
+  id="docked-buttons"
   [ngClass]="{'shadow':  !width}"
   [hidden]="!currentComponent">
   <li
