+++ core/gui/src/app/workspace/component/property-editor/property-editor.component.scss
@@ -42,7 +42,7 @@
   display: flex;
 }
 
-#property-buttons1 {
+#docked-buttons {
   position: fixed;
   top: 10vh;
   right: 0;
