+++ core/gui/src/app/workspace/component/property-editor/property-editor.component.ts
@@ -6,6 +6,7 @@ import { UntilDestroy, untilDestroyed } from "@ngneat/until-destroy";
 import { filter } from "rxjs/operators";
 import { PortPropertyEditFrameComponent } from "./port-property-edit-frame/port-property-edit-frame.component";
 import { NzResizeEvent } from "ng-zorro-antd/resizable";
+import { calculateTotalTranslate3d } from "../../../common/util/panel-dock";
 /**
  * PropertyEditorComponent is the panel that allows user to edit operator properties.
  * Depending on the highlighted operator or link, it displays OperatorPropertyEditFrameComponent
@@ -40,7 +41,7 @@ export class PropertyEditorComponent implements OnInit, OnDestroy {
     const style = localStorage.getItem("right-panel-style");
     if (style) document.getElementById("right-container")!.style.cssText = style;
     const translates = document.getElementById("right-container")!.style.transform;
-    const [xOffset, yOffset, _] = this.calculateTotalTranslate3d(translates);
+    const [xOffset, yOffset, _] = calculateTotalTranslate3d(translates);
     this.returnPosition = { x: -xOffset, y: -yOffset };
     this.registerHighlightEventsHandler();
   }
@@ -119,34 +120,5 @@ export class PropertyEditorComponent implements OnInit, OnDestroy {
 
   resetPanelPosition() {
     this.dragPosition = { x: this.returnPosition.x, y: this.returnPosition.y };
-  }
-
-  parseTranslate3d(translate3d: string): [number, number, number] {
-    const regex = /translate3d\((-?\d+\.?\d*)px,\s*(-?\d+\.?\d*)px,\s*(-?\d+\.?\d*)px\)/g;
-    const match = regex.exec(translate3d);
-    if (match) {
-      const x = parseFloat(match[1]);
-      const y = parseFloat(match[2]);
-      const z = parseFloat(match[3]);
-      return [x, y, z];
-    }
-    return [0, 0, 0];
-  }
-
-  calculateTotalTranslate3d(translates: string): [number, number, number] {
-    let totalXOffset = 0;
-    let totalYOffset = 0;
-    let totalZOffset = 0;
-
-    const translate3dArray = translates.match(/translate3d\(.*?\)/g) || [];
-
-    for (const translate of translate3dArray) {
-      const [x, y, z] = this.parseTranslate3d(translate);
-      totalXOffset += x;
-      totalYOffset += y;
-      totalZOffset += z;
-    }
-
-    return [totalXOffset, totalYOffset, totalZOffset];
   }
 }
